Update mysteryclock.ino
fixed halting bug before in the time adjust function by adding an extra call for the 'second' beforehand (line 143). This changed (shortened) the amount it was ever adjusted by and so a 2 second increment replaces it. Also if the adjust is called while the 'second' is '0' or '1' it is delayed 2 seconds to prevent adjustment errors (would go to 59 without rolling the minute back presuming it checks if it was 0 first at all (it did) or else it goes negative and may damage the rtc module).